// ClientApp/src/app/models/donor-model.ts
export class Donor{
        id: string;
        fullName: string;
        password: string;
        firstName: string;
        lastName: string;
        connectionID: string;
        ageGroup: string;
        email: string;
        phoneNumber: string;
        country: string;
        city: string;
        nativeLanguage: string;
        totalDonation: number;
        lastDonation: number;
        vip: boolean;
        anashIsrael: boolean;
        anashUSA: boolean;
        pinskSchoolGraduate: boolean;
        kievSchoolGraduate: boolean;
        yeshivaGraduate: boolean;
        inPinsk: boolean;
        businessAssociate: boolean;
        boysCounselor: boolean;
        girlsCounselor: boolean;
        helpedByPinsk: boolean;
        generalSupporter: boolean;
        mhsg: boolean;
        belarusAnsectors: boolean;
        belarusTourism: boolean;
        yyFundraiser: boolean;
        yyFamily: boolean;
        yyStaff: boolean;
        rShteiermanFamily: boolean;
        rFimaFamily: boolean;
        marriedAYYGraduate: boolean;
        yearsInYadYisroel: number;
        other: string;
    }