// ClientApp/src/app/models/member-model.ts
export class Member{
    id: string;
    fullName: string;
    password: string;
    firstName: string;
    lastName: string;
    confirmed: boolean;
    ageGroup: string;
    email: string;
    phoneNumber: string;
    country: string;
    city: string;
    fluentHebrew: boolean;
    fluentRussian: boolean;
    fluentEnglish: boolean;
    anashIsrael: boolean;
    anashUSA: boolean;
    pinskSchoolGraduate: boolean;
    kievSchoolGraduate: boolean;
    yeshivaGraduate: boolean;
    inPinsk: boolean;
    businessAssociate: boolean;
    boysCounselor: boolean;
    girlsCounselor: boolean;
    helpedByPinsk: boolean;
    generalSupporter: boolean;
    mhsg: boolean;
    belarusAnsectors: boolean;
    belarusTourism: boolean;
    yyFundraiser: boolean;
    yyFamily: boolean;
    yyStaff: boolean;
    rShteiermanFamily: boolean;
    rFimaFamily: boolean;
    marriedAYYGraduate: boolean;
    yearsInYadYisroel: number;
    yearsIAsFundraiser: number;
    operationRoom: string;
    other: string;
}